Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -6,10 +6,10 @@
 
 int main(int argc, char *const av [])
 {
-    float a = 0;
-    if ((scanf("%f \n", &a)) != 1)
+    double a = 0.0f;
+    if ((scanf("%lf \n", &a)) != 1)
         return ERROR;
-    if ((printf("You entered: %0.4f\n", a)) < 0)
+    if ((printf("You entered: %0.4lf\n", a)) < 0)
         return ERROR;
     return SUCCESS;
 }
